code: create the first working version of the program
The program retrieves the minimum duty cycle to use from the first
eeprom's address and falls back to 0.5% if unset. It supports the
TINY versions of AVR devices, with input on PB0 and output on PB1.
It simply copies the input to the output and sends small pulses
when the input is idle for too long. With high a enough clock it
doesn't introduce any measurable latency effect, as the signal
will typically be shifted by less than 10 cycles. Getting out of
idle may take up to one millisecond if the input signal arrives at
the beginning of the pulse, but even one millisecond corresponds
to only 20 microns at an engraving speed of 1200mm/min so that's
harmless.