circuit: add the laserdot schematic and board
These ones are really trivial since there are only 4 components but they can
save time to those willing to reimplement the circuit. The output uses 4 pin
at 2.0 pitch to match my NEJE-20W laser, and the input uses a male JST-XH
2.54mm connector to match the connector at the end of the cable on the laser
module side. It's worth noting that the cable is not straight but reversed,
so the polarity here really requires the cable or needs to be reversed.